// frontend/src/interfaces/Controller.ts
export interface CSS {
  css: string;
  pageId: number;
}
// For simple usage
// export type Controller = Image | Page | Label | Icon | Fader | Meter | Button | Text;
export interface Controller {
  id: number;
  type: string;
  pageId: number;
  css: string;
  cssClass: string;
  active: boolean;
  width: number | string;
  height: number | string;
  top: number | string;
  left: number | string;

  // Image properties
  srcType?: string; // Options: local | external
  value?: string;
  objectFit?: string;

  // Page properties
  targetPage?: string;
  title?: string;

  // Label and flex properties
  // Flex properties
  flexDirection?: string;
  flexWrap?: string;
  alignContent?: string;
  justifyContent?: string;
  alignItems?: string;
  // Label properties
  tag?: string;
  textAlign?: string;
  // value: 'Label', // Matches with page value

  // Icon properties
  // value: '', // Matches with page and label value

  // Fader properties
  handle?: number;
  min?: number;
  max?: number;
  vertical?: boolean;
  current_value?: number;

  // Meter properties
  // Same as fader

  // Button properties
  // handle
  toggle?: boolean;
  toggleValue?: number;
  setValue?: number | string;
  // current_value: 0
  label?: string;

  // Text properties
  operation?: string;
  // label: 'Text'
  labelPosition?: string;
  selected?: boolean;
}